myapp/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytest==9.0.2
pytest-flask==1.3.0